Добавление соединения с сервером (теперь по нормальному)
