--- Code/gui/main_app.py
@@ -1,48 +1,0 @@
-import dearpygui.dearpygui as dpg
-from dearpygui_async import DearPyGuiAsync
-from systems.loc import Localization as loc
-
-from .conntect import ConnectManager
-from .fonts import FontManager
-
-
-class DMClientApp:
-    def __init__(self):
-        self.dpg_async = DearPyGuiAsync()
-
-    def setup(self):
-        dpg.create_context()
-        FontManager.load_fonts()
-        dpg.create_viewport(title=loc.get_string("main-app-name"), width=600, height=400)
-        dpg.setup_dearpygui()
-        self.create_warning_window()
-        
-        self.con_manager = ConnectManager()
-
-    def create_warning_window(self):
-        with dpg.window(label="Warning", modal=True, tag="warning_window", no_move=True, no_close=True, no_collapse=True):
-            dpg.add_text(loc.get_string("main_text_warning_window"))
-            dpg.add_button(label=loc.get_string("yes_warning_window"), callback=self.on_yes)
-            dpg.add_button(label=loc.get_string("no_warning_window"), callback=self.on_no)
-
-    def on_yes(self, sender, app_data):
-        dpg.delete_item("warning_window")
-        self.create_main_window()
-
-    def on_no(self, sender, app_data):
-        self.stop()
-
-    def create_main_window(self):
-        with dpg.window(label="Main", tag="main_window"):
-            dpg.add_button(label=loc.get_string("connect_window_open"), callback=self.con_manager.create_connect_window)
-            dpg.add_button(label="debug", callback=lambda: dpg.show_debug())
-            dpg.add_button(label="demo", callback=lambda: dpg.show_imgui_demo())
-    
-    def run(self):
-        dpg.show_viewport()
-        self.dpg_async.run()
-        dpg.start_dearpygui()
-        dpg.destroy_context()
-
-    def stop(self):
-        dpg.stop_dearpygui()
