--- Content/Client/loc/rus/connect_window.loc
@@ -1,6 +1,0 @@
-connect_main_text = Соединение с сервером
-connect_login_hint = Логин
-connect_password_hint = Пароль
-connect_host_hint = IP
-connect_port_lable = Порт
-connect_button = Соединиться
