--- Content/Client/loc/rus/main_app.loc
@@ -1,5 +1,0 @@
-main-app-name = DM-Bot Client
-
-main_text_warning_window = Внимание! Использование DM-Bot разрешено только для пользователей, достигших 18-летнего возраста.
-yes_warning_window = Мне 18 или более лет
-no_warning_window = Мне нет 18 лет
