+++ Content/Client/loc/rus/standarts.loc
@@ -1,0 +1,3 @@
+ok = ОК
+yes = Да
+no = Нет
